finished refactoring config tests
